requirements.txt
flask==2.3.3
flask-wtf==1.2.1
openai==1.3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
gunicorn==21.2.0
diff-match-patch==20230430
pytest==7.4.3
pytest-flask==1.3.0
Werkzeug==2.3.7
Jinja2==3.1.2
Flask-SQLAlchemy==3.1.1
